[agent_builder/skills] - refactor: extract state logic into useSkillsAndActionsState hook
 - Removed useMemo-based state logic in AgentBuilderSkillsBlock.tsx to a separate hook for clarity and reusability
 - Created a new file skillsAndActionsState.ts containing the extracted logic to manage the skills and actions state
 - Simplified AgentBuilderSkillsBlock.tsx by replacing inline state computation with the useSkillsAndActionsState hook call
